% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TweedSpot-package.R
\docType{package}
\name{TweedSpot-package}
\alias{TweedSpot}
\alias{TweedSpot-package}
\title{TweedSpot: Self-Adaptive Spatial Tweedie Models for Spatially Variable Gene Detection}
\description{
TweedSpot provides spatially variable gene detection for spatial omics data
using Tweedie generalized additive models with spatial smoothing.
}
\details{
The package currently exposes \code{filter_genes()} for preprocessing and
\code{tweedspot()} for spatially variable gene detection with Tweedie generalized
additive models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mallick-lab/TweedSpot}
  \item Report bugs at \url{https://github.com/mallick-lab/TweedSpot/issues}
}

}
\author{
\strong{Maintainer}: Nalin Arora \email{naa4050@med.cornell.edu} (\href{https://orcid.org/0009-0009-1340-688X}{ORCID}) (affiliation: Cornell University)

Authors:
\itemize{
  \item Nalin Arora \email{naa4050@med.cornell.edu} (\href{https://orcid.org/0009-0009-1340-688X}{ORCID}) (affiliation: Cornell University)
  \item Himel Mallick \email{him4004@med.cornell.edu} (affiliation: Cornell University)
  \item Vivian Jia-yin Myers \email{FIRST.LAST@cornell.edu} (affiliation: Cornell University)
}

}
\keyword{internal}
