% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{top_spatial_genes}
\alias{top_spatial_genes}
\title{Return the top TweedSpot hits}
\usage{
top_spatial_genes(
  input,
  n = 10L,
  rank_by = c("padj", "pval", "stat", "dev_expl")
)
}
\arguments{
\item{input}{A \link[SpatialExperiment:SpatialExperiment]{SpatialExperiment::SpatialExperiment} containing TweedSpot
results in \code{rowData(input)}.}

\item{n}{Number of rows to return. Default \code{10}.}

\item{rank_by}{Ranking column. Options are \code{"padj"}, \code{"pval"}, \code{"stat"}, or
\code{"dev_expl"}. Default \code{"padj"}.}
}
\value{
A ranked \code{data.frame} containing the top TweedSpot hits.
}
\description{
\code{top_spatial_genes()} returns the top \code{n} rows from
\code{\link[=get_tweedspot_results]{get_tweedspot_results()}} using a user-selected ranking metric.
}
\examples{
\dontrun{
top_spatial_genes(spe_res, n = 10)
}
}
